+++ ignite/cmd/scaffold_chain_registry.go
@@ -1,0 +1,68 @@
+package ignitecmd
+
+import (
+	"github.com/spf13/cobra"
+
+	"github.com/ignite/cli/v28/ignite/pkg/cliui"
+	"github.com/ignite/cli/v28/ignite/services/chain"
+	"github.com/ignite/cli/v28/ignite/services/scaffolder"
+)
+
+// NewScaffoldChainRegistry returns the command to scaffold the chain registry chain.json and assets.json files.
+func NewScaffoldChainRegistry() *cobra.Command {
+	c := &cobra.Command{
+		Use:   "chain-registry",
+		Short: "Configs for the chain registry",
+		Long: `Scaffold the chain registry chain.json and assets.json files.
+
+The chain registry is a GitHub repo, hosted at https://github.com/cosmos/cosmos-registry, that
+contains the chain.json and assets.json files of most of chains in the Cosmos ecosystem.
+It is good practices, when creating a new chain, and about to launch a testnet or mainnet, to
+publish the chain's metadata in the chain registry.
+
+Read more about the chain.json at https://github.com/cosmos/chain-registry?tab=readme-ov-file#chainjson
+Read more about the assets.json at https://github.com/cosmos/chain-registry?tab=readme-ov-file#assetlists`,
+		Args:    cobra.NoArgs,
+		PreRunE: migrationPreRunHandler,
+		RunE:    scaffoldChainRegistryFiles,
+	}
+
+	flagSetPath(c)
+	flagSetClearCache(c)
+
+	c.Flags().AddFlagSet(flagSetYes())
+
+	return c
+}
+
+func scaffoldChainRegistryFiles(cmd *cobra.Command, _ []string) error {
+	session := cliui.New(cliui.StartSpinnerWithText(statusScaffolding))
+	defer session.End()
+
+	c, err := chain.NewWithHomeFlags(cmd)
+	if err != nil {
+		return err
+	}
+
+	cfg, err := c.Config()
+	if err != nil {
+		return err
+	}
+
+	appPath := flagGetPath(cmd)
+	sc, err := scaffolder.New(appPath)
+	if err != nil {
+		return err
+	}
+
+	if err = sc.AddChainRegistryFiles(c, cfg); err != nil {
+		return err
+	}
+
+	// no need for post scaffolding, as we are just creating two files
+	// that are not part of the build process
+
+	session.Printf("🎉 chain-registry files successfully scaffolded\n")
+
+	return nil
+}